+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master (unreleased)
 
+* Performance improvements, calculate attributes to serialize only once. ([#98](https://github.com/petalmd/bright_serializer/pull/98))
 * Add instrumentation. ([#90](https://github.com/petalmd/bright_serializer/pull/90))
 
 ## 0.3.1 (2022-09-28)